try get storage

+++ models/monitoring.go
@@ -22,8 +22,8 @@ type SysComponent struct {
 }
 
 type CompStats struct {
-	Total      int64 `json:"total"`
-	Current    int64 `json:"current"`
+	Total      float64 `json:"total"`
+	Current    float64 `json:"current"`
 	Percentage int   `json:"percentage"`
 }
